refactor(core): split api_jobs.rs god-file into modules
