+++ src/core/api_jobs/mod.rs
@@ -1,1012 +1,32 @@
 #![allow(dead_code)]
 
-use std::collections::HashMap;
-use std::fmt;
-use std::fs;
-use std::path::{Path, PathBuf};
-use std::sync::atomic::{AtomicU64, Ordering};
-use std::sync::{Arc, Mutex};
-use std::thread::{self, JoinHandle};
-use std::time::{SystemTime, UNIX_EPOCH};
-
-use serde::{Deserialize, Serialize};
-use serde_json::Value;
-use uuid::Uuid;
-
-use crate::core::error::{Error, Result};
-use crate::core::redaction::redact_argv_display;
-use crate::core::source_snapshot::SourceSnapshot;
-
+mod persistence;
 mod remote_runner;
+mod store;
+mod summary;
+mod types;
+
 pub use remote_runner::{
     JobArtifactMetadata, RemoteRunnerJobClaim, RemoteRunnerJobRequest, RemoteRunnerJobResult,
 };
-
-const DEFAULT_EVENT_RETENTION_LIMIT: usize = 1000;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum JobStatus {
-    Queued,
-    Running,
-    Succeeded,
-    Failed,
-    Cancelled,
-}
-
-impl JobStatus {
-    fn is_terminal(self) -> bool {
-        matches!(self, Self::Succeeded | Self::Failed | Self::Cancelled)
-    }
-
-    pub fn run_status_label(self) -> &'static str {
-        match self {
-            Self::Queued => "queued",
-            Self::Running => "running",
-            Self::Succeeded => "pass",
-            Self::Failed => "fail",
-            Self::Cancelled => "cancelled",
-        }
-    }
-
-    pub fn daemon_status_label(self) -> &'static str {
-        match self {
-            Self::Queued => "queued",
-            Self::Running => "running",
-            Self::Succeeded => "succeeded",
-            Self::Failed => "failed",
-            Self::Cancelled => "cancelled",
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum JobEventKind {
-    Status,
-    Stdout,
-    Stderr,
-    Progress,
-    Result,
-    Error,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Job {
-    pub id: Uuid,
-    pub operation: String,
-    pub status: JobStatus,
-    pub created_at_ms: u64,
-    pub updated_at_ms: u64,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub started_at_ms: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub finished_at_ms: Option<u64>,
-    pub event_count: usize,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub source_snapshot: Option<SourceSnapshot>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub stale_reason: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub target_runner_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub target_project_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub claim_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub claimed_by_runner_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub claimed_at_ms: Option<u64>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub claim_expires_at_ms: Option<u64>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub artifacts: Vec<JobArtifactMetadata>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ActiveRunnerJobSummary {
-    pub runner_id: String,
-    pub job_id: String,
-    pub operation: String,
-    pub source: String,
-    pub kind: String,
-    pub status: JobStatus,
-    pub command: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub cwd: Option<String>,
-    pub started_at_ms: u64,
-    #[serde(default)]
-    pub updated_at_ms: u64,
-    pub elapsed_ms: u64,
-    #[serde(default)]
-    pub heartbeat_age_ms: u64,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub claim_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub claimed_by_runner_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub claimed_at_ms: Option<u64>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub claim_expires_at_ms: Option<u64>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub claim_expires_in_ms: Option<u64>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub durable_run_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub stale_reason: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub lifecycle_state: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub retryable: Option<bool>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub active_child_count: Option<u64>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub active_cell_count: Option<u64>,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum RunnerJobSource {
-    Broker,
-    Daemon,
-    #[serde(rename = "runner-daemon")]
-    RunnerDaemon,
-    #[serde(rename = "direct-daemon")]
-    DirectDaemon,
-    #[serde(rename = "reverse-broker")]
-    ReverseBroker,
-    Unknown,
-}
-
-impl RunnerJobSource {
-    pub fn label(self) -> &'static str {
-        match self {
-            Self::Broker => "broker",
-            Self::Daemon => "daemon",
-            Self::RunnerDaemon => "runner-daemon",
-            Self::DirectDaemon => "direct-daemon",
-            Self::ReverseBroker => "reverse-broker",
-            Self::Unknown => "unknown",
-        }
-    }
-
-    pub fn from_metadata(value: &str) -> Self {
-        match value {
-            "broker" => Self::Broker,
-            "daemon" => Self::Daemon,
-            "runner-daemon" | "runner_daemon" => Self::RunnerDaemon,
-            "direct-daemon" | "direct_daemon" => Self::DirectDaemon,
-            "reverse-broker" | "reverse_broker" => Self::ReverseBroker,
-            _ => Self::Unknown,
-        }
-    }
-
-    pub fn lifecycle_owner(self) -> RunnerJobLifecycleOwner {
-        match self {
-            Self::Broker | Self::ReverseBroker => RunnerJobLifecycleOwner::Broker,
-            _ => RunnerJobLifecycleOwner::Controller,
-        }
-    }
-}
-
-impl fmt::Display for RunnerJobSource {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        formatter.write_str(self.label())
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum RunnerJobLifecycleOwner {
-    Broker,
-    Controller,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ActiveRunnerJobRunSummary {
-    pub id: String,
-    pub kind: String,
-    pub status: String,
-    pub started_at: String,
-    pub command: String,
-    pub cwd: Option<String>,
-    pub status_note: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct JobEvent {
-    pub sequence: u64,
-    pub job_id: Uuid,
-    pub kind: JobEventKind,
-    pub timestamp_ms: u64,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub message: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub data: Option<Value>,
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct JobStore {
-    inner: Arc<Mutex<JobStoreInner>>,
-    next_event_sequence: Arc<AtomicU64>,
-    persistence: Option<Arc<JobStorePersistence>>,
-}
-
-#[derive(Debug, Clone)]
-struct JobStorePersistence {
-    path: PathBuf,
-    event_retention_limit: usize,
-}
-
-#[derive(Debug, Default)]
-struct JobStoreInner {
-    jobs: HashMap<Uuid, StoredJob>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct StoredJob {
-    job: Job,
-    events: Vec<JobEvent>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    remote_runner: Option<remote_runner::StoredRemoteRunnerJob>,
-}
-
-#[derive(Debug, Default, Serialize, Deserialize)]
-struct DurableJobStore {
-    jobs: Vec<StoredJob>,
-}
-
-#[derive(Debug)]
-pub struct JobRunner {
-    pub job_id: Uuid,
-    pub handle: JoinHandle<()>,
-}
-
-#[derive(Debug, Clone)]
-pub struct JobHandle {
-    store: JobStore,
-    job_id: Uuid,
-}
-
-impl JobStore {
-    pub(crate) fn open(path: impl Into<PathBuf>) -> Result<Self> {
-        Self::open_with_event_retention(path, DEFAULT_EVENT_RETENTION_LIMIT)
-    }
-
-    pub(crate) fn open_with_event_retention(
-        path: impl Into<PathBuf>,
-        event_retention_limit: usize,
-    ) -> Result<Self> {
-        let path = path.into();
-        let mut durable = read_durable_store(&path)?;
-        let event_retention_limit = event_retention_limit.max(1);
-        let next_sequence = reconcile_stale_jobs(&mut durable, event_retention_limit);
-        let store = Self {
-            inner: Arc::new(Mutex::new(JobStoreInner {
-                jobs: durable
-                    .jobs
-                    .into_iter()
-                    .map(|stored| (stored.job.id, stored))
-                    .collect(),
-            })),
-            next_event_sequence: Arc::new(AtomicU64::new(next_sequence)),
-            persistence: Some(Arc::new(JobStorePersistence {
-                path,
-                event_retention_limit,
-            })),
-        };
-
-        store.persist()?;
-        Ok(store)
-    }
-
-    pub(crate) fn create(&self, operation: impl Into<String>) -> Job {
-        self.create_with_source_snapshot(operation, None)
-    }
-
-    pub(crate) fn create_with_source_snapshot(
-        &self,
-        operation: impl Into<String>,
-        source_snapshot: Option<SourceSnapshot>,
-    ) -> Job {
-        let now = timestamp_ms();
-        let job = Job {
-            id: Uuid::new_v4(),
-            operation: operation.into(),
-            status: JobStatus::Queued,
-            created_at_ms: now,
-            updated_at_ms: now,
-            started_at_ms: None,
-            finished_at_ms: None,
-            event_count: 0,
-            source_snapshot,
-            stale_reason: None,
-            target_runner_id: None,
-            target_project_id: None,
-            claim_id: None,
-            claimed_by_runner_id: None,
-            claimed_at_ms: None,
-            claim_expires_at_ms: None,
-            artifacts: Vec::new(),
-        };
-
-        let mut inner = self.inner.lock().expect("job store mutex poisoned");
-        inner.jobs.insert(
-            job.id,
-            StoredJob {
-                job: job.clone(),
-                events: Vec::new(),
-                remote_runner: None,
-            },
-        );
-        drop(inner);
-
-        self.append_status_event(job.id, JobStatus::Queued, "job queued")
-            .expect("newly-created job must accept queued status event");
-        self.get(job.id)
-            .expect("newly-created job must be readable after insert")
-    }
-
-    pub(crate) fn get(&self, job_id: Uuid) -> Result<Job> {
-        let inner = self.inner.lock().expect("job store mutex poisoned");
-        let stored = inner
-            .jobs
-            .get(&job_id)
-            .ok_or_else(|| job_not_found(job_id))?;
-        Ok(stored.job.clone())
-    }
-
-    pub(crate) fn list(&self) -> Vec<Job> {
-        let inner = self.inner.lock().expect("job store mutex poisoned");
-        let mut jobs: Vec<Job> = inner
-            .jobs
-            .values()
-            .map(|stored| stored.job.clone())
-            .collect();
-        jobs.sort_by_key(|job| (job.created_at_ms, job.id));
-        jobs
-    }
-
-    pub(crate) fn active_runner_jobs(&self) -> Vec<ActiveRunnerJobSummary> {
-        let now = timestamp_ms();
-        let inner = self.inner.lock().expect("job store mutex poisoned");
-        let mut jobs: Vec<ActiveRunnerJobSummary> = inner
-            .jobs
-            .values()
-            .filter(|stored| matches!(stored.job.status, JobStatus::Queued | JobStatus::Running))
-            .filter_map(|stored| {
-                let request = stored.remote_runner.as_ref()?.request.clone();
-                Some(active_runner_job_summary(&stored.job, &request, now))
-            })
-            .collect();
-        jobs.sort_by_key(|job| (job.started_at_ms, job.job_id.clone()));
-        jobs
-    }
-
-    pub(crate) fn stale_runner_jobs(&self) -> Vec<ActiveRunnerJobSummary> {
-        let now = timestamp_ms();
-        let inner = self.inner.lock().expect("job store mutex poisoned");
-        let mut jobs: Vec<ActiveRunnerJobSummary> = inner
-            .jobs
-            .values()
-            .filter(|stored| {
-                stored.job.status == JobStatus::Failed && stored.job.stale_reason.is_some()
-            })
-            .filter_map(|stored| {
-                let request = stored.remote_runner.as_ref()?.request.clone();
-                Some(active_runner_job_summary(&stored.job, &request, now))
-            })
-            .collect();
-        jobs.sort_by_key(|job| (job.updated_at_ms, job.job_id.clone()));
-        jobs
-    }
-
-    pub(crate) fn events(&self, job_id: Uuid) -> Result<Vec<JobEvent>> {
-        let inner = self.inner.lock().expect("job store mutex poisoned");
-        let stored = inner
-            .jobs
-            .get(&job_id)
-            .ok_or_else(|| job_not_found(job_id))?;
-        Ok(stored.events.clone())
-    }
-
-    pub(crate) fn start(&self, job_id: Uuid) -> Result<Job> {
-        self.transition(job_id, JobStatus::Running, "job started")
-    }
-
-    pub(crate) fn complete(&self, job_id: Uuid, result: Option<Value>) -> Result<Job> {
-        self.ensure_transition(job_id, JobStatus::Succeeded)?;
-        if let Some(data) = result {
-            self.append_event(job_id, JobEventKind::Result, None, Some(data))?;
-        }
-        self.transition(job_id, JobStatus::Succeeded, "job succeeded")
-    }
-
-    pub(crate) fn fail(&self, job_id: Uuid, error: impl Into<String>) -> Result<Job> {
-        self.ensure_transition(job_id, JobStatus::Failed)?;
-        let error = error.into();
-        self.append_event(job_id, JobEventKind::Error, Some(error.clone()), None)?;
-        self.transition(job_id, JobStatus::Failed, error)
-    }
-
-    pub(crate) fn cancel(&self, job_id: Uuid, reason: impl Into<String>) -> Result<Job> {
-        self.transition(job_id, JobStatus::Cancelled, reason.into())
-    }
-
-    pub(crate) fn append_event(
-        &self,
-        job_id: Uuid,
-        kind: JobEventKind,
-        message: Option<String>,
-        data: Option<Value>,
-    ) -> Result<JobEvent> {
-        let mut inner = self.inner.lock().expect("job store mutex poisoned");
-        let stored = inner
-            .jobs
-            .get_mut(&job_id)
-            .ok_or_else(|| job_not_found(job_id))?;
-        if kind != JobEventKind::Status && stored.job.status.is_terminal() {
-            return Err(Error::validation_invalid_argument(
-                "status",
-                format!("cannot append {:?} event to terminal job", kind),
-                Some(job_id.to_string()),
-                None,
-            ));
-        }
-
-        let event = JobEvent {
-            sequence: self.next_event_sequence.fetch_add(1, Ordering::SeqCst) + 1,
-            job_id,
-            kind,
-            timestamp_ms: timestamp_ms(),
-            message,
-            data,
-        };
-
-        stored.events.push(event.clone());
-        apply_event_retention(&mut stored.events, self.event_retention_limit());
-        stored.job.event_count = stored.events.len();
-        stored.job.updated_at_ms = event.timestamp_ms;
-        drop(inner);
-
-        self.persist()?;
-
-        Ok(event)
-    }
-
-    pub(crate) fn run_background<T, F>(&self, operation: impl Into<String>, run: F) -> JobRunner
-    where
-        T: Serialize + Send + 'static,
-        F: FnOnce(JobHandle) -> Result<T> + Send + 'static,
-    {
-        self.run_background_with_source_snapshot(operation, None, run)
-    }
-
-    pub(crate) fn run_background_with_source_snapshot<T, F>(
-        &self,
-        operation: impl Into<String>,
-        source_snapshot: Option<SourceSnapshot>,
-        run: F,
-    ) -> JobRunner
-    where
-        T: Serialize + Send + 'static,
-        F: FnOnce(JobHandle) -> Result<T> + Send + 'static,
-    {
-        let job = self.create_with_source_snapshot(operation, source_snapshot);
-        let job_id = job.id;
-        let handle_store = self.clone();
-        let worker_store = self.clone();
-
-        let handle = thread::spawn(move || {
-            if worker_store.start(job_id).is_err() {
-                return;
-            }
-            let job_handle = JobHandle {
-                store: handle_store,
-                job_id,
-            };
-
-            match run(job_handle) {
-                Ok(output) => {
-                    let result = serde_json::to_value(output).ok();
-                    let _ = worker_store.complete(job_id, result);
-                }
-                Err(err) => {
-                    let _ = worker_store.fail(job_id, err.to_string());
-                }
-            }
-        });
-
-        JobRunner { job_id, handle }
-    }
-
-    fn transition(
-        &self,
-        job_id: Uuid,
-        next_status: JobStatus,
-        message: impl Into<String>,
-    ) -> Result<Job> {
-        let message = message.into();
-        {
-            let mut inner = self.inner.lock().expect("job store mutex poisoned");
-            let stored = inner
-                .jobs
-                .get_mut(&job_id)
-                .ok_or_else(|| job_not_found(job_id))?;
-            validate_transition(stored.job.status, next_status)?;
-
-            let now = timestamp_ms();
-            stored.job.status = next_status;
-            stored.job.updated_at_ms = now;
-            if next_status == JobStatus::Running {
-                stored.job.started_at_ms = Some(now);
-            }
-            if next_status.is_terminal() {
-                stored.job.finished_at_ms = Some(now);
-            }
-        }
-
-        self.persist()?;
-
-        self.append_status_event(job_id, next_status, message)?;
-        self.get(job_id)
-    }
-
-    fn ensure_transition(&self, job_id: Uuid, next_status: JobStatus) -> Result<()> {
-        let inner = self.inner.lock().expect("job store mutex poisoned");
-        let stored = inner
-            .jobs
-            .get(&job_id)
-            .ok_or_else(|| job_not_found(job_id))?;
-        validate_transition(stored.job.status, next_status)
-    }
-
-    fn append_status_event(
-        &self,
-        job_id: Uuid,
-        status: JobStatus,
-        message: impl Into<String>,
-    ) -> Result<JobEvent> {
-        self.append_event(
-            job_id,
-            JobEventKind::Status,
-            Some(message.into()),
-            Some(serde_json::json!({ "status": status })),
-        )
-    }
-
-    fn event_retention_limit(&self) -> usize {
-        self.persistence
-            .as_ref()
-            .map(|persistence| persistence.event_retention_limit)
-            .unwrap_or(usize::MAX)
-    }
-
-    fn persist(&self) -> Result<()> {
-        let Some(persistence) = &self.persistence else {
-            return Ok(());
-        };
-
-        let durable = {
-            let inner = self.inner.lock().expect("job store mutex poisoned");
-            DurableJobStore {
-                jobs: inner.jobs.values().cloned().collect(),
-            }
-        };
-
-        write_durable_store(&persistence.path, &durable)
-    }
-}
-
-fn active_runner_job_summary(
-    job: &Job,
-    request: &remote_runner::RemoteRunnerJobRequest,
-    now_ms: u64,
-) -> ActiveRunnerJobSummary {
-    let started_at_ms = job.started_at_ms.unwrap_or(job.created_at_ms);
-    ActiveRunnerJobSummary {
-        runner_id: request.runner_id.clone(),
-        job_id: job.id.to_string(),
-        operation: job.operation.clone(),
-        source: request_metadata_string(request, "source")
-            .unwrap_or_else(|| "runner-daemon".to_string()),
-        kind: request_metadata_string(request, "kind").unwrap_or_else(|| job.operation.clone()),
-        status: job.status,
-        command: redact_argv_display(&request.command),
-        cwd: request.cwd.clone(),
-        started_at_ms,
-        updated_at_ms: job.updated_at_ms,
-        elapsed_ms: now_ms.saturating_sub(started_at_ms),
-        heartbeat_age_ms: now_ms.saturating_sub(job.updated_at_ms),
-        claim_id: job.claim_id.clone(),
-        claimed_by_runner_id: job.claimed_by_runner_id.clone(),
-        claimed_at_ms: job.claimed_at_ms,
-        claim_expires_at_ms: job.claim_expires_at_ms,
-        claim_expires_in_ms: job
-            .claim_expires_at_ms
-            .map(|expires_at| expires_at.saturating_sub(now_ms)),
-        durable_run_id: request_metadata_string(request, "durable_run_id")
-            .or_else(|| request_metadata_string(request, "run_id"))
-            .or_else(|| request_metadata_string(request, "record_run_id")),
-        stale_reason: job.stale_reason.clone(),
-        lifecycle_state: Some(runner_job_lifecycle_state(job).to_string()),
-        retryable: Some(runner_job_retryable(job)),
-        active_child_count: request
-            .metadata
-            .as_ref()
-            .and_then(|metadata| metadata.get("active_child_count"))
-            .and_then(Value::as_u64),
-        active_cell_count: request
-            .metadata
-            .as_ref()
-            .and_then(|metadata| metadata.get("active_cell_count"))
-            .and_then(Value::as_u64),
-    }
-}
-
-pub fn active_runner_job_run_summary(job: ActiveRunnerJobSummary) -> ActiveRunnerJobRunSummary {
-    let active_child_count = optional_count(job.active_child_count);
-    let active_cell_count = optional_count(job.active_cell_count);
-    let durable_run_id = job.durable_run_id.as_deref().unwrap_or("unknown");
-    let command = format!(
-        "{} [source={}, kind={}, runner={}, job={}, durable_run={}, elapsed_ms={}, active_child_count={}, active_cell_count={}]",
-        job.command,
-        job.source,
-        job.kind,
-        job.runner_id,
-        job.job_id,
-        durable_run_id,
-        job.elapsed_ms,
-        active_child_count,
-        active_cell_count
-    );
-    let status_note = format!(
-        "active runner job: source={} kind={} runner={} job={} durable_run={} elapsed_ms={} active_child_count={} active_cell_count={}",
-        job.source,
-        job.kind,
-        job.runner_id,
-        job.job_id,
-        durable_run_id,
-        job.elapsed_ms,
-        active_child_count,
-        active_cell_count
-    );
-
-    ActiveRunnerJobRunSummary {
-        id: job
-            .durable_run_id
-            .clone()
-            .unwrap_or_else(|| format!("runner-job-{}", job.job_id)),
-        kind: job.kind,
-        status: job.status.run_status_label().to_string(),
-        started_at: ms_to_rfc3339(job.started_at_ms),
-        command,
-        cwd: job.cwd,
-        status_note,
-    }
-}
-
-fn optional_count(count: Option<u64>) -> String {
-    count
-        .map(|count| count.to_string())
-        .unwrap_or_else(|| "unknown".to_string())
-}
-
-fn ms_to_rfc3339(ms: u64) -> String {
-    chrono::DateTime::<chrono::Utc>::from_timestamp_millis(ms as i64)
-        .unwrap_or_else(chrono::Utc::now)
-        .to_rfc3339()
-}
-
-fn runner_job_lifecycle_state(job: &Job) -> &'static str {
-    if job.status == JobStatus::Failed
-        && job.stale_reason.as_deref()
-            == Some("daemon restarted before the job reached a terminal status")
-    {
-        "abandoned_after_daemon_restart"
-    } else if job.stale_reason.is_some() {
-        "stale"
-    } else if matches!(job.status, JobStatus::Queued | JobStatus::Running) {
-        "active"
-    } else {
-        "terminal"
-    }
-}
-
-fn runner_job_retryable(job: &Job) -> bool {
-    matches!(
-        runner_job_lifecycle_state(job),
-        "abandoned_after_daemon_restart" | "stale"
-    )
-}
-
-fn request_metadata_string(
-    request: &remote_runner::RemoteRunnerJobRequest,
-    key: &str,
-) -> Option<String> {
-    request
-        .metadata
-        .as_ref()
-        .and_then(|metadata| metadata.get(key))
-        .and_then(Value::as_str)
-        .map(str::to_string)
-}
-
-fn read_durable_store(path: &Path) -> Result<DurableJobStore> {
-    if !path.exists() {
-        return Ok(DurableJobStore::default());
-    }
-
-    let content = fs::read_to_string(path)
-        .map_err(|e| Error::internal_io(e.to_string(), Some(format!("read {}", path.display()))))?;
-    match serde_json::from_str(&content) {
-        Ok(store) => Ok(store),
-        Err(err) => {
-            let quarantine_path = path.with_file_name(format!(
-                "{}.corrupt-{}",
-                path.file_name()
-                    .and_then(|value| value.to_str())
-                    .unwrap_or("jobs.json"),
-                timestamp_ms()
-            ));
-            fs::rename(path, &quarantine_path).map_err(|rename_err| {
-                Error::config_invalid_json(path.display().to_string(), err).with_hint(format!(
-                    "Homeboy could not quarantine the corrupt durable job store to {}: {}",
-                    quarantine_path.display(),
-                    rename_err
-                ))
-            })?;
-            eprintln!(
-                "Homeboy quarantined corrupt daemon job store {} to {} and started with an empty queue",
-                path.display(),
-                quarantine_path.display()
-            );
-            Ok(DurableJobStore::default())
-        }
-    }
-}
-
-fn write_durable_store(path: &Path, durable: &DurableJobStore) -> Result<()> {
-    if let Some(parent) = path.parent() {
-        fs::create_dir_all(parent).map_err(|e| {
-            Error::internal_io(e.to_string(), Some(format!("create {}", parent.display())))
-        })?;
-    }
-
-    let body = serde_json::to_string_pretty(durable).map_err(|e| {
-        Error::internal_json(
-            e.to_string(),
-            Some("serialize daemon job store".to_string()),
-        )
-    })?;
-    fs::write(path, body)
-        .map_err(|e| Error::internal_io(e.to_string(), Some(format!("write {}", path.display()))))
-}
-
-fn reconcile_stale_jobs(durable: &mut DurableJobStore, event_retention_limit: usize) -> u64 {
-    let now = timestamp_ms();
-    let mut next_sequence = durable
-        .jobs
-        .iter()
-        .flat_map(|stored| stored.events.iter().map(|event| event.sequence))
-        .max()
-        .unwrap_or(0);
-
-    for stored in &mut durable.jobs {
-        if !matches!(stored.job.status, JobStatus::Queued | JobStatus::Running) {
-            continue;
-        }
-        // Remote-runner jobs that are still Queued are waiting for a runner to
-        // claim them; a daemon restart does not invalidate that work unless the
-        // non-serialized execution request carried secret env values.
-        if stored.remote_runner.is_some() && stored.job.status == JobStatus::Queued {
-            if !remote_runner_job_has_unrecoverable_execution_env(stored) {
-                continue;
-            }
-        }
-
-        // Recover the real terminal status when the underlying command already
-        // recorded a terminal Result event before the daemon restarted. Without
-        // this, a job that actually succeeded (or that recorded its own
-        // non-zero exit code) is blindly reported as a daemon-restart failure,
-        // leaving the caller without the real result for in-flight work (#4770).
-        if let Some((recovered_status, exit_code)) = recovered_terminal_from_result(&stored.events)
-        {
-            stored.job.status = recovered_status;
-            stored.job.updated_at_ms = now;
-            stored.job.finished_at_ms = Some(now);
-            stored.job.stale_reason = None;
-
-            next_sequence += 1;
-            stored.events.push(JobEvent {
-                sequence: next_sequence,
-                job_id: stored.job.id,
-                kind: JobEventKind::Status,
-                timestamp_ms: now,
-                message: Some(
-                    "job terminal status recovered from recorded result after daemon restart"
-                        .to_string(),
-                ),
-                data: Some(serde_json::json!({
-                    "status": recovered_status,
-                    "reason": "recovered_after_daemon_restart",
-                    "exit_code": exit_code,
-                })),
-            });
-            apply_event_retention(&mut stored.events, event_retention_limit);
-            stored.job.event_count = stored.events.len();
-            continue;
-        }
-
-        let reason = if remote_runner_job_has_unrecoverable_execution_env(stored) {
-            "daemon restarted before the remote runner claimed secret execution env".to_string()
-        } else {
-            "daemon restarted before the job reached a terminal status".to_string()
-        };
-        stored.job.status = JobStatus::Failed;
-        stored.job.updated_at_ms = now;
-        stored.job.finished_at_ms = Some(now);
-        stored.job.stale_reason = Some(reason.clone());
-
-        next_sequence += 1;
-        stored.events.push(JobEvent {
-            sequence: next_sequence,
-            job_id: stored.job.id,
-            kind: JobEventKind::Error,
-            timestamp_ms: now,
-            message: Some(reason.clone()),
-            data: Some(serde_json::json!({ "reason": "stale_after_daemon_restart" })),
-        });
-        next_sequence += 1;
-        stored.events.push(JobEvent {
-            sequence: next_sequence,
-            job_id: stored.job.id,
-            kind: JobEventKind::Status,
-            timestamp_ms: now,
-            message: Some("job marked failed after daemon restart".to_string()),
-            data: Some(serde_json::json!({
-                "status": JobStatus::Failed,
-                "reason": "stale_after_daemon_restart"
-            })),
-        });
-        apply_event_retention(&mut stored.events, event_retention_limit);
-        stored.job.event_count = stored.events.len();
-    }
-
-    next_sequence
-}
-
-fn remote_runner_job_has_unrecoverable_execution_env(stored: &StoredJob) -> bool {
-    let Some(remote_runner) = stored.remote_runner.as_ref() else {
-        return false;
-    };
-    if remote_runner.execution_request.is_some() {
-        return false;
-    }
-    remote_runner.request.secret_env_names.iter().any(|name| {
-        remote_runner
-            .request
-            .env
-            .get(name)
-            .is_some_and(|value| value == "<redacted>")
-    })
-}
-
-/// Recover a terminal job status from a recorded `Result` event when a job was
-/// left non-terminal by a daemon restart. The daemon worker records the command
-/// result (including its `exit_code`) before transitioning the job to its
-/// terminal status; if the restart lands in that window the stored result is the
-/// authoritative outcome. Returns the recovered status and the exit code that
-/// justified it, or `None` when no terminal result was recorded.
-fn recovered_terminal_from_result(events: &[JobEvent]) -> Option<(JobStatus, i64)> {
-    let result = events
-        .iter()
-        .rev()
-        .find(|event| event.kind == JobEventKind::Result)?;
-    let data = result.data.as_ref()?;
-    // A recorded cancellation outcome is honored as Cancelled regardless of exit code.
-    if data.get("status").and_then(Value::as_str) == Some("cancelled") {
-        return Some((
-            JobStatus::Cancelled,
-            data.get("exit_code").and_then(Value::as_i64).unwrap_or(0),
-        ));
-    }
-    let exit_code = data.get("exit_code").and_then(Value::as_i64)?;
-    let status = if exit_code == 0 {
-        JobStatus::Succeeded
-    } else {
-        JobStatus::Failed
-    };
-    Some((status, exit_code))
-}
-
-fn apply_event_retention(events: &mut Vec<JobEvent>, limit: usize) {
-    if events.len() > limit {
-        let excess = events.len() - limit;
-        events.drain(0..excess);
-    }
-}
-
-impl JobHandle {
-    pub(crate) fn job_id(&self) -> Uuid {
-        self.job_id
-    }
-
-    pub(crate) fn is_cancelled(&self) -> bool {
-        self.store
-            .get(self.job_id)
-            .map(|job| job.status == JobStatus::Cancelled)
-            .unwrap_or(true)
-    }
-
-    pub(crate) fn stdout(&self, message: impl Into<String>) -> Result<JobEvent> {
-        self.store.append_event(
-            self.job_id,
-            JobEventKind::Stdout,
-            Some(message.into()),
-            None,
-        )
-    }
-
-    pub(crate) fn stderr(&self, message: impl Into<String>) -> Result<JobEvent> {
-        self.store.append_event(
-            self.job_id,
-            JobEventKind::Stderr,
-            Some(message.into()),
-            None,
-        )
-    }
-
-    pub(crate) fn progress(&self, data: Value) -> Result<JobEvent> {
-        self.store
-            .append_event(self.job_id, JobEventKind::Progress, None, Some(data))
-    }
-
-    pub(crate) fn result(&self, data: Value) -> Result<JobEvent> {
-        self.store
-            .append_event(self.job_id, JobEventKind::Result, None, Some(data))
-    }
-}
-
-fn validate_transition(current: JobStatus, next: JobStatus) -> Result<()> {
-    let allowed = matches!(
-        (current, next),
-        (JobStatus::Queued, JobStatus::Running)
-            | (JobStatus::Queued, JobStatus::Cancelled)
-            | (JobStatus::Running, JobStatus::Succeeded)
-            | (JobStatus::Running, JobStatus::Failed)
-            | (JobStatus::Running, JobStatus::Cancelled)
-    );
-
-    if allowed {
-        Ok(())
-    } else {
-        Err(Error::validation_invalid_argument(
-            "status",
-            format!("cannot transition job from {:?} to {:?}", current, next),
-            None,
-            None,
-        ))
-    }
-}
-
-fn job_not_found(job_id: Uuid) -> Error {
-    Error::validation_invalid_argument("job_id", "job not found", Some(job_id.to_string()), None)
-}
-
-fn timestamp_ms() -> u64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .expect("system clock must be after unix epoch")
-        .as_millis() as u64
-}
+pub use store::{JobHandle, JobRunner, JobStore};
+pub use summary::active_runner_job_run_summary;
+pub use types::{
+    ActiveRunnerJobRunSummary, ActiveRunnerJobSummary, Job, JobEvent, JobEventKind, JobStatus,
+    RunnerJobLifecycleOwner, RunnerJobSource,
+};
 
 #[cfg(test)]
 mod tests {
     use std::collections::HashMap;
+    use std::fs;
 
     use serde_json::json;
 
+    use super::persistence::recovered_terminal_from_result;
     use super::*;
+    use crate::core::source_snapshot::SourceSnapshot;
+    use uuid::Uuid;
 
     #[test]
     fn test_create() {
@@ -1305,7 +325,7 @@ mod tests {
     fn background_runner_captures_errors_as_failed_jobs() {
         let store = JobStore::default();
         let runner = store.run_background::<serde_json::Value, _>("test", |_job| {
-            Err(Error::validation_invalid_argument(
+            Err(crate::core::error::Error::validation_invalid_argument(
                 "fixture", "boom", None, None,
             ))
         });
@@ -1993,7 +1013,7 @@ mod tests {
                 .get_mut(&job.id)
                 .expect("job exists")
                 .job
-                .claim_expires_at_ms = Some(timestamp_ms().saturating_sub(1));
+                .claim_expires_at_ms = Some(super::persistence::timestamp_ms().saturating_sub(1));
         }
 
         let expired_event = store.append_remote_runner_event(
@@ -2056,7 +1076,7 @@ mod tests {
                 .get_mut(&job.id)
                 .expect("job exists")
                 .job
-                .claim_expires_at_ms = Some(timestamp_ms().saturating_sub(1));
+                .claim_expires_at_ms = Some(super::persistence::timestamp_ms().saturating_sub(1));
         }
 
         let expired = store.renew_remote_runner_claim(job.id, "homeboy-lab", &claim_id, 30_000);
